feat!: use FilteredReadExec in the planner (#3813)
Switches the planner to use the new FilteredReadExec node when not using
legacy storage.

Starts to remove the storage_class prototype. Keeping it up was too
difficult in this change and the prototype has not seen use or active
development into a proper feature. That can still happen in the future.

Refactors most of the planner to use `Projection` instead of `Schema` to
represent the user's projection.

BREAKING CHANGE: The names `_rowid` and `_rowaddr` are no longer valid
column names (it's unclear they ever should have been and using these
names would have likely led to errors)