fix: fix errors caused by simultaneous merge (#4279)
Two commits merged recently and conflicted with each other.


https://github.com/lancedb/lance/commit/59eff93d2cb0f3a75a9c5cecf58cc7a3c2b8b797
changes some merge_insert logic and uses parse_filter in the planner

https://github.com/lancedb/lance/commit/eb5e1c4279d63634b9e99ca48a14425fc72f554d
changed some planner logic and added a check to ensure the filter in
parse_filter is a boolean

Unfortunately, that check required looking at the schema and some of the
filters in the new merge_insert path looked like `source.xyz` and that
confused the check. For now I've moved the check out into the scanner
instead of the planner.