refactor: use previous instead v2 for our old encoding algo (#4336)
We are currently working on the `v2.1` format, but many of our code
paths are still using `v2`, which is actually the previous
implementation. This PR renames the old legacy v2 code to `previous` to
make it clearer.

I find this useful because it can prevent LLMs from referring to or
searching within incorrect code paths.

Signed-off-by: Xuanwo <github@xuanwo.io>