refactor: simplify substrait logic, allow session state to be injected (#4357)
Parsing and encoding Substrait requires a Datafusion `SessionState` in
order to look up UDFs. I was working on an integration that uses
Datafusion UDFs and it was failing to parse Substrait filters because
the parse routine was using a newly created default session context.
This PR makes the session state configurable. It also gets rid of the
old "dummy plan" approach and encodes the extended expression message
directly.