fix: make sure cloud features aren't implicitly enabled (lancedb/lancedb#2567)
lance-io includes cloud features and critical types/methods that sibling
libraries need without strictly needing cloud features. Make sure that
lance-io is depended on by default without the default feature set
except for lance which will pull in the requisite features as they're
enabled.

This is so that building lancedb with default-features = false excludes
the entire dependency chain for crates supporting the cloud features,
shrinking the size and speeding up the build of the purely embedded
version.