fix: don't panic when reading files if the last row was deleted (#4396)
The new filtered read applies deletion vectors before the read. There
was a bug in this logic so that, if the last row in the fragment was
deleted, an empty range would be sent to the file reader. For example,
if a file had 50 rows and row 49 was deleted we would ask for rows
`[0..49], [50..50]`.

The file reader then had a bug that, when it processed that empty range,
it would panic. This seems specific to 2.0 files with the list data type
though the fix is to just ignore empty ranges at the top level.