feat: allow empty projection (#4340)
Recent changes made it so we didn't report an error in `project` if a
column did not exist. This made it possible to get an error "no columns
selected" when all columns asked for did not exist. This was quite
confusing.

This PR cleans up that behavior by reworking the projection slightly.

Previously, if a user requested zero columns, then they would receive an
error "no columns selected and with_row_id is False, there is nothing to
scan". With this PR we now return empty record batches (with the correct
number of rows).