feat: allow configuring bss for encoding (#4393)
This PR enables users to configure BSS for encoding.

We have implemented an internal BSS sensitivity concept to measure how
sensitive we want to be, based on our byte position entity. However, we
will not expose these details directly to users. Instead, we provide a
simple interface for users to choose between `on`, `off`, and `auto`.
The default value is `auto`, which means BSS will be enabled based on
the data's statistics.

We can introduce additional statistics functions in the future. I don't
see any blockers or conflicts.

---------

Signed-off-by: Xuanwo <github@xuanwo.io>