refactor: refactor encoding proto for 2.1 to cleanup and future-proof (#4434)
This makes a full scan through the 2.1 encoding schemes and cleans up
the protobuf. This removes a number of schemes that were 2.0 only and
helps cleanup the intent. In addition, many schemes are hard-coding an
inner encoding of flat today. We might want to do something different in
the future. This changes the protos so the hard-coding is done in code
and not protobuf. This should help future-proof the encoding scheme for
minor changes in 2.2 or 2.1.2 or whatever is next.