fix: none ASCII path issue (#4442)
Close #4021

I encountered this issue in another PR: #4257. In that PR I need to
persist a dataset url as String and use the String to re-construct a
dataset path. I've stuck in a '~' encoding issue for some time(different
with the original Chinese character issue, but I think the same encoding
problem), which drives me to figure out the url/path encoding things.

Here's the comparision:

- Path::from()will encode both the non ASCII strings like Chinese
characters and the reserved ASCII characters like ~
- Url::parse() will encode only non ASCII strings like Chinese
characters
- Path::parse() will encode **nothing**
- Path::from_url_path will decode from a path

I think the key is If we want to generate a path from a url, we should
assume the encoding done or not(or let's say the url is invalid or not,
not encoded means invalid). Most cases:
1. We get a url by Url::parse(), which has encoded the non ASCII
strings(if the string has been encoded, it won't encode again)
2. If we use a local path to generate a url, the internal string has
been encoded.

If we assume the url is trusted in encoding, then we should always use
Path::parse(). I think this is what we should do. If the path is
invalid, meaning the url or the original string should be encoded first.
This will avoid double encoding and unnecessary decoding in the
Path::from_url_path case.

Note:
1. Use Url::parse and Path::parse will **NOT** encode the reserved ASCII
characters like ~,?,&, for now it works well. The object_store path
could handle the reserved ASCII characters correctly(refer to aws.rs
testsutes).
2. For % in original paths(like 'aaa/bbb/cc%cc'), the % will **NOT** be
encoded. We need to use urlencoding or other libs encode the url/path
string before it is used. （seems quite reasonable to me)
3. Local paths seem more complicated. Modern OS supports non ASCII
paths, but Windows and Linux/Mac seems different on the path processing.
I think keeping the principle would be fine.

Please take a look when you have time @jackye1995 @Xuanwo @ddupg cc
@wjones127 @yanghua