test: add recurring test (#4190)
We've seen some bugs that could be reproduced with some operations in
some order, so adding this for finding out such bugs:
- https://github.com/lancedb/lance/pull/4108
- https://github.com/lancedb/lance/pull/4156
- https://github.com/lancedb/lance/pull/3769

These weekly tests would enumerate all permutations of write operations,
and run all of them, to make sure we can cover all cases. And we would
run these tests on latest 2 stable version and main, to find out any
regression and compatibility issues

This would take a long time, so we run these weekly, rather than on each
commit.

and we've seen some bugs which happened on only large dataset, we can
run some tests over a large dataset weekly, but not adding these for now

---------

Signed-off-by: BubbleCal <bubble-cal@outlook.com>