fix: the elapsed_time is not correct for IvfSubIndex Node (#4491)
noticed that the elapsed time is much higher than the time I measured,
and it's generally equal to `real_elapsed * nprobes`.

this is because the elapsed time is the sum of each partition.

I don't have a good idea to verify this by unit test, so locally tested
it and verified the fix works

---------

Signed-off-by: BubbleCal <bubble-cal@outlook.com>