fix: skip rebasing for strict overwrite (#4377)
The "strict overwrite" commit type, which respects `read_version` and
skips all retries for Overwrite operations when `num_retries=0`, is
broken at the Python level by the automatic conflict resolution.

The rebasing is checking out the most recent version of the dataset,
causing `read_version` to be ignored.

This PR fixes the issue by skipping automatic rebasing when in strict
overwrite mode, and adds a Python test for the behavior.

---------

Co-authored-by: Weston Pace <weston.pace@gmail.com>