feat: allow reading the _rowoffset (#4478)
The row offset (_rowoffset) is similar to the row id and the row address
but slightly different. The row offset is always a number from 0 to
len(dataset).

The row offset is highly unstable (it can change from rows being deleted
or from compaction, updates, or anything else that modifies a dataset)
but it has one useful property. The user does not need to query the list
of available row offset up front for sampling purposes. This makes it
useful in some cases for sampling a dataset.

In addition, this PR reworks the way we do projection. It generalizes
things like _rowaddr and _rowid as "system columns". We can now read
_rowid and _rowaddr without requiring the user to call with_row_id or
with_row_address. If they are specified in a projection then they will
show up in the output in the order in which they were projected.
However, if the user calls with_row_id or with_row_offset then we will
still always place them at at the end.

In addition, scoring columns (e.g. `_distance`, `_score`) can be
specified in the projection. To avoid breaking changes we still include
those columns even if they are not included in a projection. However,
this behavior will change in the future. This PR adds a warning that
will be emitted and a new flag (`disable_scoring_autoprojection`) that
can be used to opt in to the new behavior early to avoid the upcoming
breaking change.