test(java): re-construct tests for merge operation (#4483)
I found merge operation test is not right. It didn't construct by
merging a new column, but instead merged some new rows. Although it
works, it's not what we wanted.

In the meanwhile the TransactionTest has grown a little tedious and not
that friendly to construct multiple testcases on a single operation. So
I did a little refactoring.

The work includes:
1. Move all operation testcases into package
com/lancedb/lance/operation.
2. Extend some new testcase for Append and Merge operation. We could do
similar work to other operations.
3. Original TransactionTest is only used to test transaction framework
like readTransaction and transaction properties.