fix: restore behavior of projection plans from schema (#4562)
In the past when we created a projection plan from a schema it was
possible to mask out certain columns while still maintaining the overall
structure. This was only exposed via rust (no way to project with schema
via python) but we (LanceDB) found at least one workflow relying on this
behavior.

The behavior changed (unintentionally) in
https://github.com/lancedb/lance/commit/eeea03c21e9c5082a1699c5eee3c14dd3d0ef60a
and this PR restores the old behavior. That being said, I don't
particularly love that there are two different ways to specify a
projection and it might be nice to update downstream workflows at some
point or come up with some way to do masking via projection expressions.