feat(java): add deleteRows for Fragment (#4528)
This PR adds a method for Dataset to delete some rows (identified by
_rowaddr list) from a specific Fragment.

This method will be used in Spark MOR update. The updated old rows are
deleted from the Fragment using Deletion Vector. And updated new rows
are written in new Fragments. In commit stage, Update transaction
operation is committed on the Dataset to apply updation.

Co-authored-by: fangbo.0511 <fangbo.0511@bytedance.com>