feat: add a scalar index for JSON (#4621)
Adds a scalar index for JSON columns by creating a scalar index on some
sub-field in the JSON data. Any scalar index type can be used as the
underlying type.

This PR also introduces the ability to configure scalar indexes in a
generic way. It might be nice to eventually switch the Inverted index so
that it can also be trained in this way (and then we don't need a
special "create full text index" method) but that can be left for future
work.

This PR also modifies the expression parsing slightly to introduce the
idea that a query might match a direct column reference _or_ might match
a projected column reference. This could be useful in the future for
things like indexes on struct fields, list items, image metadata, etc.

---------

Signed-off-by: Xuanwo <github@xuanwo.io>
Co-authored-by: Xuanwo <github@xuanwo.io>