fix: correctly decode lists with all nulls (#4679)
If all items in a list are null it triggers the complex null layout
where we write the rep/def into a buffer but nothing else. When reading
from this layout we were not handling invisible items incorrectly. As a
result, a list with all nulls would still read out having items and
return a null items array. This could lead to errors if the list's items
field was non-nullable.