feat: create indexes from another source dataset (#4658)
Closes #4613 

This PR adds the ability for a dataset to initialize its indices from
another source dataset. This requires these 2 datasets to have the same
schema at least for the indexed columns (same column name and field ID).
For vector index specifically, the index will also inherit the centroids
from the source vector index, so that the index in this dataset is
basically a delta index of the source dataset. From the context of
MemTable work, the MemTable is another Lance table, and we bootstrap
this table by initializing it against the source table using this new
feature.