fix: add incremental metrics in FilteredRead (#4798)
The existing implementation for IO metrics and scan metrics didn't work
if we stopped reading early, such as when using Limit. The IO metrics
were only recorded at the end of the stream, so would be left at zero if
the stream was terminated early. The scan metrics were reported during
the planning stage, so they always reported the full scan metrics, even
if we only read a small portion of the data.

Fixes #4791