feat: implement blob encoding for 2.1 (#4802)
This PR fully adds the blob encoding back into 2.1. The PR is slightly
lazy at the moment. For scheduling we simply load the descriptions as
part of the search cache in the initialization of the scheduler. In the
future we may not want to do that since it could be memory intensive. We
should instead spawn a new task to decode the descriptions via indirect
I/O. That can be saved for a follow-up however.