feat(java): expose additional Operation::Update fields to bindings (#4788)
This PR enhances the Java bindings for Operation::Update as a follow-up
to #4589.

While #4589 introduced new fields to the Rust Operation::Update struct,
the Java bindings were still missing support not only for those new
fields but also for the existing `fields_modified field`, which is
crucial for handling column modifications.

Specifically, it adds support for the following fields:

```rust
  Update {
      // ...
      /// The fields that have been modified
      fields_modified: Vec<u32>,
      // 👇 New fields from #4589
      /// The fields that used to judge whether to preserve the new frag's id into
      /// the frag bitmap of the specified indices.
      fields_for_preserving_frag_bitmap: Vec<u32>,
      /// The mode of update
      update_mode: Option<UpdateMode>,
  }
```

---------

Co-authored-by: Weiren <litaiwei.lwt@antgroup.com>