refactor: cowardly refuse to use scalar indexes on expressions with null literals (#4815)
The current search logic with scalar indexes assumes that the list
returned by a scalar index contains either:

* All the true rows and all other rows are false
* All the false rows and all other rows are true

This is a problem when dealing with nulls. This is related to
https://github.com/lancedb/lance/issues/4756

This PR doesn't fix #4756 however it does help avoid some areas where we
are definitely likely to do the wrong thing.