feat: upgrade url crate, rework temporary directory utils (#4860)
There is a [behavior
change](https://github.com/servo/rust-url/issues/1077) in url 2.5.7 that
is blocking a datafusion upgrade. This PR works around that behavior
change. Unfortunately, the fix here is to just avoid problematic Windows
paths in our tests while we wait for guidance from upstream. It will
still be possible for users to encounter errors when providing Windows
paths. We can look at hacks for that problem in a future PR (or just
hope the upstream bug is fixed promptly).