feat: support fragment-level update columns (#4715)
As described in https://github.com/lancedb/lance/issues/4650, this PR
implements a fragment-level column update operation in Rust and exposes
it through the Java API. This allows for efficient, partial updates of a
dataset without copying irrelevant fields.

The update mechanism follows these rules:
* **Update existing rows**: For any row identified by a given key in the
target table, if a corresponding update value is present in the source
data, the target row's value is replaced. This applies even if the new
value is null.
* **Preserve untouched rows**: If a row in the target table has no
corresponding update in the source data, it remains unchanged.

The image below illustrates this process:
<img width="536" height="289" alt="image"
src="https://github.com/user-attachments/assets/d3629f88-e3f2-4f9d-b0e9-379d3dd482d1"
/>


**Important:** It is the responsibility of the calling engine (e.g.,
Spark, Flink) to ensure that the update data (the source) is correctly
partitioned and routed to the corresponding fragments of the target
table. The core operation assumes it receives the correct data for the
fragment it is operating on.

---------

Co-authored-by: Weiren <litaiwei.lwt@antgroup.com>