fix: filter out vectors that could produce infinite l2 distances (#4890)
This is a quick fix for #4842.

I think a long-term fix will need to make this aware of the distance
type. Or maybe we filter when actually computing distances later, so we
don't have false-positives. I know this mostly applies for l2. Cosine
(IIRC) normalizes the vectors and doesn't have this problem.