fix: short circuit SQL parsing (#4825)
This PR avoids a problem encountered when creating a scalar index on a
column that had a name that was not easily parseable by the SQL parser.
If we are getting a raw expression that matches a column name in the
schema, then return it directly.

More generally I wonder if we should be doing this sql parsing at all
when we know we have columns. Should `ProjectionPlan::from_expressions`
actually take in instead `columns: &[(impl AsRef<str>, &Expr)]` ?