fix: forward compatibility of FTS index (#4906)
When creating using 0.38 and reading using 0.36 for `list_indices` the
operation fails with

```
thread '<unnamed>' panicked at /home/runner/work/lance/lance/rust/lance-index/src/scalar/inverted/index.rs:846:30:
called `Option::unwrap()` on a `None` value
```

This PR bumps the version of the FTS index when the token set format is FTS, so that the incompatible index should be ignored by old clients. However, there is a bug in the old client that if the indices are cached when opening manifest, then the index is not ignored and would still cause error when doing list_indices or other operations against the index.