feat: handle forking better (#4903)
When a process which is making use of lance calls fork(), threads in that process will not carry over into
the child process. This PR tries to handle the situation slightly better by re-creating thread executors
in the child process. Note that we still do not want to officially support fork. For one thing, tokio has
some issues with it. Therefore, we will still print a warning and recommend using a non-forking mode
for multiprocessing.