fix: rewrap LanceFilterExec with_new_children (#4920)
Prior to this commit, calling with_new_children on a LanceFilterExec
would produce a FilterExec physical plan node, from which the filter
could not be recovered in optimizer rules.

This caused an issue in our optimizer rules after a recent DF update -
presumably this is now getting called somewhere where it previously was
not.

This commit wraps the output in a new LanceFilterExec so that the filter
is still available to optimizer rules.