fix: rebuild HNSW graph while remapping it (#4941)
HNSW can't be simply remapped because the graph structure also needs to
be rebuilt after the vectors changed,
we didn't allow to remap HNSW before this, this PR allows remapping HNSW
by rebuilding the graph.

Before this, compaction on a table with HNSW index would raise error
"remap HNSW is not supported yet". Users would be able to do remap
(triggered by compaction) after this.

---------

Signed-off-by: BubbleCal <bubble-cal@outlook.com>