fix: clarify column names in missing column errors (#4942)
Prior to this commit we could fail to find expected internal columns in
KNN execution and display an unprefixed name to the user. This is
confusing because there's no visual indicator that the column is
internal (indicated by prefixing with _ typically). This commit changes
the error to display the real column name instead.