test: fix flaky commit iops test (#4970)
The test checks the number of iops with and without a cache. However, it
sets the cache size to 10,000. My guess is this was originally done when
the cache size was a number of items. Now it is being interpreted as
10KB. Some of the transactions in the test are a few KB. So the test
actually always broke the cache limit.

The test was only flaky (and typically only failed on Windows) because
moka does not evict immediately (if it did the test would fail every
time). So Windows, being slower, is more likely to give enough time for
moka to run an eviction pass, which would cause an occasional failure.