refactor!: cleanup public API, remove lance_arrow re-exports (#4991)
I'd like to start addressing
https://github.com/lancedb/lance/issues/3864 so that we can reduce the
amount of breaking changes we introduce by minimizing the public API
surface that we commit to. Regrettably, doing so involves creating many
breaking changes :face_exhaling: as we have too many exports today as
well as duplicate exports.

This PR is the first in a series of PRs to cleanup the public API. All
of these changes are breaking changes.

* Reduces the scope of lance_arrow re-exports to just the bfloat16
array. The rest of the lance_arrow consists of convenience utilities,
which are handy, but not something we want to commit to supporting
indefinitely, and things like FloatType, which we might re-export as
part of lance-index, but is not generally applicable for lance users.
* Removes the re-export of lance_core::error. We already re-export
`Error` and `Result` and the rest of the stuff in there is not meant for
public consumption
* ~~Removes the `crate::arrow::json` utilities. These converted data
types and schemas to/from JSON. This is neat, but only used in lancedb
today. It should be included there, or put in its own crate. We don't
want to commit to maintaining this.~~
* Change `crate::session::index_extension` to be `pub(crate)`
* Removes lance_datafusion::sql re-export. This was a convenience export
from an earlier refactor that is no longer needed.
* Removes lance_index::vector::kmeans for the same reason