fix: ensure limit cancels scan (#5032)
The hard limit in filtered_read was being applied with a try_filter_map.
Unfortunately, this operator will not "quit" in the same way that a
take_while does so this PR adds a take_while to ensure the stream quits
scanning once it sees enough rows to satisfy the limit.