refactor: use lance-io object store for dir namespace and improve builder (#5045)
I ended up doing these in https://github.com/lancedb/lance/pull/4984 and
https://github.com/lancedb/lance/pull/4905 so I decided to pull it out
and get it cleaned up first.

This PR moves the directory namespace from using OpenDAL directly to
using Lance ObjectStore. This avoids the inconsistency between the dir
namespace and the underlying lance table storage configurations. User
can still use OpenDAL, and if we fully migrate Lance to OpenDAL it will
be applied to both layers at the same time as well.

The PR also improves the builder of the namespaces with builder style
and allow supplying a Lance session. Since we have not published a
stable version yet, we do not care about backwards compatibility.

This PR also ensures the lance-namespace-impls features are consistent
with lance-io features. Related to
https://github.com/lancedb/lance/issues/5042