refactor: don't add new with_... columns to scanner (#5007)
We recently added `with_row_last_updated_at_version` and
`with_row_created_at_version` methods to `Scanner`. I'd like to avoid
that, preferring to use `project` to ask for those columns. This is
because we may have quite a few system columns over time and I don't
want the scanner to have lots of `with_` methods. These methods need to
be added to a lot of places (python bindings, take methods, java
bindings, lancedb, etc.) In addition, the `with_` methods make it
unclear the expected output order of the columns (e.g. if I specify
`with_row_id` _and_ `with_row_addr` then which comes first).

This refactor changes things so we just use `project`.

However, this introduced one small wrinkle, which is that we need to be
able to say "all columns plus these system columns". I introduced the
concept of `*` to allow this so you can no do something like
`project(&["*", "_rowid"])`.

I need to add more test cases for the wildcard but wanted to verify that
this is an approach everyone is happy with first.