feat: support dynamic storage options provider with AWS credentials vending (#4905)
This PR introduces a new dynamic storage options provider interface in
Lance dataset. The main idea is that the provider describes the storage
options to use for a given dataset, and when these options will expire.
Lance is responsible for fetching another new set of storage options to
re-initialize the object store when expiration happens.

This is mainly useful for cases where the dataset's access credentials
are temporary, and the user would like to invoke a specific credentials
endpoint to fetch a new set of credentials. Currently we have only added
support for AWS by implementing a credentials provider.

The PR also provides an implementation of the `StorageOptionsProvider`
with Lance Namespace, because Lance Namespace provides a DescribeTable
endpoint which returns the storage options that should be used at the
given time. Based on the namespace spec, a `expires_at_millis` key can
be added to the storage options to indicate the expiration time of those
options.

Because Lance Namespace provides native implementations in python and
Java, we also provides binding interfaces of of
`PyStorageOptionsProvider` and `JavaStorageOptionsProvider`, which then
further integrated with Lance Namespace implemented in those specific
languages.