fix: ensure I/O cancels correctly when scan is dropped (#5129)
Previously the scheduler took a "poison the well approach". When the
scheduler was dropped, it would poison any remaining tasks so they
returned an error.

Unfortunately, this approach is not easily plugged into Datafusion,
which does not have any kind of asynchronous cancellation of the stream.

Instead, Datafusion encourages a "abort all tasks" approach when the
stream is cancelled. This PR migrates things to the "abort all tasks"
approach. It also uses SpawnedTask::spawn in filtered_read instead of
tokio::task::spawn to create abort-on-drop fire and forget tasks.

In addition, this PR connects the `io_buffer_size` property to filtered
read. This scanner property was previously being ignored. I needed it
for a unit test.