fix: prevent cache collisions in ObjectStoreRegistry (#5153)
ObjectStoreRegistry maintains a cache of ObjectStore objects.
Previously, when caching an Azure blob store, the account name was not
part of the cache key. This could result in collisions in the case where
we tried to cache two Azure blob stores using the same container name,
but different account names. Container names are not unique in Azure;
only the combination of account name, container name is.

The flow here is roughly that we determine what type of object store
we're creating by looking at the first part of the URL (the scheme), and
then use the associated object store Provider object to generate the
appropriate store_prefix from the second part of the URL (the
authority). We want to be able to do all this relatively quickly,
without instantiating heavyweight objects, to support the cache key
use-case.

In the specific case of Azure, the account name may not even be in the
URL at all. It can be passed as part of the storage options. Therefore,
the function generating the object store prefix must also take the
storage_options as a parameter. The final object store prefix we
generate for Azure must include this account information, as well as the
container information.

This PR modifies WrappingObjectStore.wrap() to take the store prefix as
an argument, rather than the storage options. In general, the storage
options are highly specific to the type of the object store, which the
wrapper generally doesn't know. It would not make sense to look for
Azure-specific storage options when wrapping an s3 store, for example.