feat: late materialization of vectors in filtered vector search
KNN search is performed when a vector index is not present. When a table
is partially covered by a vector index, we perform a union of an ANN
search over the indexed data, and a KNN search over the unindexed data.
If the table is completely unindexed it is just a KNN search on the
data.

Prior to this commit, when we would execute the KNN portion of a
filtered vector search, we would perform a scan of all columns and
remove results that did not match the filter. For large vectors, this
amounts to a lot of overfetch from storage.

When filters are selective, it is more efficient to read the filter
column (typically much smaller than the vector), apply the filter, and
then select matching vectors by row ID.

This patch implements that strategy as well as an adaptive mechanism for
deciding when to apply it. There is a new configuration concept in the
scanner for specifying the filter selectivity at which it will be
cheaper to do a scan. We will compute a target rowcount based on that
threshold and scan the filter column for matches. If we encounter more
matches than the target, we will give up and switch to a scan.