feat: add adapter for REST namespace with manifest namespace backend (#4984)
This PR introduces an adapter (using Axum) that provides a REST server
which delegates requests to be fulfilled by another Lance namespace.
This is helpful for easy setup of a REST namespace server to perform
testing.

In addition, I added a namespace implementation backed by a Lance table,
basically the same thing as
https://github.com/lancedb/lancedb/pull/2708, also have
DirectoryNamespace able to use it in a compatibility mode.

I moved it here as a namespace implementation, because (1) this allows
us to fully test tables in a namespace with storage only requirement,
and at the same time, (2) comparing to offer this in lancedb directly,
we know we want to consolidate the local listing connection and remote
connection as just 2 types of namespace connections, and this gives user
an incentive to move to the namespace-based connection in lancedb for
the local one

Also add corresponding python binding for DirectoryNamespace and
RestNamespace, deprecate the current ones in lance-namespace python
package