ci: cache rust dependencies, enable parallelism (1.9x to 5.5x faster) (#5236)
In #4139, we stopped caching built dependencies. Since then, we've had
to wait 20 minutes for the test to build, before we even run any tests.

This PR re-enables the build cache and implements a different mitigation
to avoid large caches: we disable debug info in dependencies and only
keep it in our crates. This is inspired by DataFusion's CI:
https://github.com/apache/datafusion/blob/377c0fce481d561d58e3a6fad2dca18cb1d58384/Cargo.toml#L247-L259

Below are the reductions in size of the target directory. Previously,
Rust jobs were running with `debuginfo=1`, so the target directory was
about 16GB. With these changes, it's down to 7.6GB. If we decided later
we wanted it even smaller, we could enable `-C opt-level=s`, which gets
us down to 2.1GB.

| Profile | Flags | `target/` size | % reduction |
|--------|--------|-------:|------:|
| dev |  | 25.0GB | - |
| dev | `-C debuginfo=1` | 16.0GB | 36% |
| dev | `-C debuginfo=line-tables-only` | 12.0GB | 52% | 
| ci | | 7.6GB | 70% | 
| ci | `-C opt-level=s` | 2.1GB | 92% |

Other optimizations added:

* Removed `CARGO_BUILD_JOBS`, which was preventing parallelism in
builds. We can add this back to specific runners if we find it's
necessary. This will speed up cold builds.
* Align profile across all builds in CI.

Thanks to reduction in disk space, closes #5218.

### Performance improvement

Most of the speed improvement probably comes from removing
`CARGO_BUILD_JOBS`. Jobs that run tests aren't much faster since the
tests take a while.

| Case | Before | After (cold) | mul | After (cached) | mul |
|--------|-------:|---:|---:|---:|---:|
| linux-arm | 38m 58s |  20m 5s | **1.9x** | 18m 33s | **2.1x** |
| MSRV Check | 11m 46s | 5m 17s | **2.3x** | 2m 8s | **5.5x** |
| Clippy | 13m 16s | 5m 26s | **2.4x** | 2m 25s | **5.4x** |